+++ CHANGELOG.md
@@ -17,6 +17,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 
 
+
+
+## v2.29.3 - 2026-02-09
+
+### Added
+- added separate setting to hide info messages (so you can hide errors, but still have quest prog messages)
+## v2.29.2 - 2026-02-09
+
+### Changed
+- disable castbar previews on profile change, this should fix the perma castbar preview issue (happened, when profile settings got copied with previews on)
 ## v2.29.1 - 2026-02-09
 
 ### Fixed